# Improve canonicalization performance with a faster `lmsr` function