This PR replaces the `lmsr` function with a significantly faster approach, adapted from an external source. The performance improvement is particularly evident with longer sequences.

I benchmarked the functions using `hyperfine`, processing three sequences with an average length of 208,399 characters.

```
Benchmark 1: old
  Time (mean ± σ):     357.1 ms ±   4.6 ms    [User: 272.9 ms, System: 91.6 ms]
  Range (min … max):   354.2 ms … 362.4 ms    3 runs

Benchmark 2: new
  Time (mean ± σ):     34.582 s ±  1.003 s    [User: 34.283 s, System: 0.216 s]
  Range (min … max):   33.657 s … 35.648 s    3 runs

Summary
  new ran
   96.84 ± 3.07 times faster than old
```